be/int/web/registration/reception: moves down.
