+++ backend/cmd/sessions/main.go
@@ -4,13 +4,14 @@ import (
 	"context"
 	"fmt"
 	"log"
+
 	registry "logbook/cmd/registry/client"
 	"logbook/cmd/sessions/app"
 	"logbook/cmd/sessions/endpoints"
 	"logbook/cmd/sessions/service"
 	"logbook/internal/startup"
 	"logbook/internal/web/balancer"
-	"logbook/internal/web/reception"
+	"logbook/internal/web/registration/reception"
 	"logbook/internal/web/registryfile"
 	"logbook/internal/web/router"
 	"logbook/internal/web/sidecar"
