+++ backend/cmd/tags/main.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	"log"
+
 	registry "logbook/cmd/registry/client"
 	sessions "logbook/cmd/sessions/client"
 	"logbook/cmd/tags/app"
@@ -11,7 +12,7 @@ import (
 	"logbook/cmd/tags/service"
 	"logbook/internal/startup"
 	"logbook/internal/web/balancer"
-	"logbook/internal/web/reception"
+	"logbook/internal/web/registration/reception"
 	"logbook/internal/web/registryfile"
 	"logbook/internal/web/router"
 	"logbook/internal/web/sidecar"
