+++ backend/cmd/users/main.go
@@ -4,13 +4,14 @@ import (
 	"context"
 	"fmt"
 	"log"
+
 	"logbook/cmd/profiles/app"
 	"logbook/cmd/profiles/endpoints"
 	"logbook/cmd/profiles/service"
 	registry "logbook/cmd/registry/client"
 	"logbook/internal/startup"
 	"logbook/internal/web/balancer"
-	"logbook/internal/web/reception"
+	"logbook/internal/web/registration/reception"
 	"logbook/internal/web/registryfile"
 	"logbook/internal/web/router"
 	"logbook/internal/web/sidecar"
